adding dark theme
Generalization of css values via variables allow to toggle between different themes by setting a global data attribute.
This toggling is done with few JS that saves the set theme to the session storage of the browser. If not manually set, it detects the system's theme and adapts it.
A toggle button is placed on the top right of every page, for this the default theme template is modified to incorporate the back-home button.