fixing theme issue
In some browser setups, the css variables could not be evaluated. This may have been caused by a css selector that detects the users system theme in order to match it even when visiting the website for the first time. As browser compatibility was not always ensured and the behaviour was very weird, it is now removed to test.
Test was successful.