adding tag feature
Adding a template that is searching for tags in blog posts and generates a list of all tags along with posts that include these. Tags are displayed at the beginning of each post where they are clickable and direct to their position on the tag list.
Linked to blog archive.