blog subdirectory
Moving all blog pages to the subdirectory /blog/ in order to keep the page organized and clearly distinct blog posts from other non-blog pages. In order to be (slightly) backwards compatible for links that were shared and opened frequently, a redirect page is added for those blog posts which use a new layout template. The blog archive stayed the same but its link was changed to /blog to allow users who currently are on a post to move the url tree up and land in the archive instead of a 404.