fixing subpage of blog
The timeline viewer linked in this blog entry should be on the same base path as the blog itself. As in commit d058fd190c0072ea9bc85e83ed373c97f45dd081 all blog entries were moved into the blog/ directory, the viewer page also had to be moved into this dir. Problem here is that this blog dir is generated by jekyll in the first place, I created it manually with only the viewer inside which could confuse. The link from the blog should work now instead of showing a 404.