search feature
Adding new feature that allows the visitor to search through the entire website and a page dedicated for it.
A .json template collects all the relevant data, a script by Fei can search through. The results are rendered on the search page as a simple list.
Also adding a button next to the theme switcher to quickly get to the search page from everywhere.