fixing reCAPTCHA issue
Targetting issue #26 which caused a non-filled reCAPTCHA form to get accepted just fine. This resulted in lots of spam comments.
Implementing a random math question captcha that doesn't reveal the answer anywhere to fetch. The question itself is displayed in FIGlet and should be hard to read by a bot.
Removing every bit of code that was used for Googles reCAPTCHA API.