restructuring attractor gallery
The first project page now has a preview gallery. Splitting the previous list of attractors into separate pages which can be navigated like the blog posts.
For page generation, details are stored in yaml format in the frontmatter of the attractor pages.
Moving wallpapers and previews into a separate folder to make it more organised.