typographical changes
Changed the main font from JetBrains Mono to iA Quattro, added an accent with Alegreya for titles. In total, that is still 66KB less data (JetBrains is huge with lots of additional symbols)! Also added an initial dropped cap for the first paragraph in blog posts and other sites with more text. To fit the overall change, the text dividers were also modified and all 3rd party icons are now removed!